Move FragmentStashViewModel into a file of it's own type
